unnesessry static warning eliminated. Many bugs fixed.
git-svn-id: http://co19.googlecode.com/svn/trunk/co19/tests/co19@584 2ecb45c1-dfb3-dc95-330a-13f216d23a04

+++ src/LibTest/core/Duration/operator_div_A01_t01.dart
@@ -18,7 +18,7 @@ void check(int d, int h, int m, int s, int ms, num q) {
       ).inMilliseconds);
 }
 
-long seed=1234567;
+int seed=1234567;
 int nextRand() {
   seed = seed * 199933 + 11;
   return seed&127;
